requirements.txt
fastapi[standard]==0.115.13
numpy==2.1.3
Pillow==11.2.1
pydantic==2.11.7
tensorflow==2.19.0
